Fixing range check in call_indirect

+++ Wacs.Core/Instructions/Control.cs
@@ -857,12 +857,12 @@ namespace Wacs.Core.Instructions
             context.Assert( context.OpStack.Peek().IsI32,
                 $"Instruction {Op.GetMnemonic()} failed. Wrong type on stack.");
             //9.
-            int i = context.OpStack.PopI32();
+            uint i = context.OpStack.PopI32();
             //10.
             if (i >= tab.Elements.Count)
                 throw new TrapException($"Instruction call_indirect could not find element {i}");
             //11.
-            var r = tab.Elements[i];
+            var r = tab.Elements[(int)i];
             //12.
             if (r.IsNullRef)
                 throw new TrapException($"Instruction call_indirect NullReference.");
